pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "flake8",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[tool.flake8]
max-line-length = 100
extend-ignore = "E203, W503"
exclude = [".git", "__pycache__", ".venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=gendiff --cov-report=term-missing"

[tool.coverage.run]
source = ["gendiff"]
omit = ["*/tests/*", "*/test_*"]

[dependency-groups]
dev = [
    "autopep8>=2.0.0",
    "pytest-cov>=5.0.0",
]
